backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
supabase
simple-salesforce
openai
pinecone-client
python-multipart
python-jose[cryptography]
passlib[bcrypt]
httpx
python-dotenv
